+++ components/form-field-fileinput/src/defaultArgs.ts
@@ -2,7 +2,6 @@ import { IFileInputFieldProps } from './template';
 import { defaultArgs as fieldDefaultArgs } from '../../form-field/src/defaultArgs';
 import { defaultArgs as fileInputDefaultArgs } from '../../form-fileinput/src/defaultArgs';
 
-export const defaultArgs: IFileInputFieldProps = {
-  ...fieldDefaultArgs,
-  ...fileInputDefaultArgs,
-};
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+const { fieldId, ...restFieldArgs } = fieldDefaultArgs;
+export const defaultArgs: IFileInputFieldProps = { ...restFieldArgs, ...fileInputDefaultArgs };