+++ components/form-textarea/CHANGELOG.md
@@ -1,5 +1,17 @@
 # `@nl-rvo/css-form-textarea`
 
+## 1.2.4
+
+### Patch Changes
+
+- 0157ba8: Text Area classes where not properly set in scss file
+
+## 1.2.3
+
+### Patch Changes
+
+- 22e4bed: Allign textarea remaining character to the right, instead of the left
+
 ## 1.2.2
 
 ### Patch Changes